cli: Call TriggerCloseCli.php instead of TriggerCloseApi.php
Every cli call in Mantis which loaded the plugin automatically fired the
closing process. It's not only bad practice, it might be hard to
diagnose as well.

Instead, we decouple the calling in a cheap way by creating a new,
by the plugin unreferenced, file to trigger the closing. The better way
would be to move all cli-related code in the API to this file, but I'm
lazy for now.

Reported-by: Félix Rauch <felix.rauch@gmail.com>
Signed-off-by: Carl Helmertz <helmertz@gmail.com>